backend/requirements.txt
flask
python-dotenv
requests
pillow
pytest
google-generativeai
